@@ -5,22 +5,25 @@
     "homepage": "https://7oskaaa.github.io/Ahmed-Hossam/",
     "dependencies": {
         "@coreui/bootstrap-react": "^1.0.0-beta.0",
-        "@emailjs/browser": "^3.10.0",
-        "@emotion/react": "^11.10.5",
-        "@emotion/styled": "^11.10.5",
-        "@mui/material": "^5.11.2",
-        "@ramonak/react-progress-bar": "^5.0.3",
-        "bootstrap": "^5.2.3",
-        "jquery": "^3.6.3",
-        "react": "^16.14.0",
+        "@emailjs/browser": "^4.4.1",
+        "@emotion/react": "^11.13.3",
+        "@emotion/styled": "^11.13.0",
+        "@fortawesome/fontawesome-svg-core": "^6.6.0",
+        "@fortawesome/free-solid-svg-icons": "^6.6.0",
+        "@fortawesome/react-fontawesome": "^0.2.2",
+        "@mui/material": "^6.0.1",
+        "@ramonak/react-progress-bar": "^5.2.0",
+        "bootstrap": "^5.3.3",
+        "jquery": "^3.7.1",
+        "react": "^18.3.1",
         "react-active-menu": "^0.0.14",
         "react-alert": "^7.0.3",
-        "react-bootstrap": "^2.7.0",
-        "react-dom": "^16.14.0",
-        "react-router-dom": "^6.6.1",
+        "react-bootstrap": "^2.10.4",
+        "react-dom": "^18.3.1",
+        "react-router-dom": "^6.26.1",
         "react-scripts": "^5.0.1",
-        "react-typed": "^1.2.0",
-        "update-browserslist-db": "^1.0.11"
+        "react-typed": "^2.0.12",
+        "update-browserslist-db": "^1.1.0"
     },
     "scripts": {
         "predeploy": "npm run build",
@@ -40,6 +43,7 @@
         "not op_mini all"
     ],
     "devDependencies": {
+        "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
         "gh-pages": "^6.1.1"
     }
 }
