@@ -1,10 +1,20 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
 import './index.css';
+
 import App from './App';
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+
 // import * as serviceWorker from './serviceWorker';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+// Create a root element
+const root = ReactDOM.createRoot(document.getElementById('root'));
+
+// Render the App component inside the root element
+root.render(
+    <React.StrictMode>
+      <App />
+    </React.StrictMode>
+);
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
